Refactor classes to remove singleton behaviour
FileSystem and MockFileSystem were singletons and were created through
the FileSystemFactory in FileInfo's constructor. This was a convenient
way to switch the real file system code for the mock code.

However, the fact that they had global state was interfering with the
tests. Each test had to be set up with either the real file system or
mock version. If this was forgotten for some reason then the tests
would be compromised.

The plan now is to inject the file system class in an undocumented
constructor of FileInfo.